Fix bug where id lists were failing to fetch
We have logic to prevent requesting id lists at a frequency higher than once per minute to match the sdk. The problem is that this logic should be applied at a per sdk key granularity, but there was a bug that was causing it to be applied at a global granularity.

This could cause subsequent sdk key get_id_list requests to skip an http request.

The fix is the properly hash the logic by sdk key.